Сделал main.js модулем, создал модули: data.js(для создания данных) и utils.js(для вспомогательных функций)
